milestones-page is not abstract
