+++ frontend/crd/src/app/app.module.ts
@@ -15,7 +15,6 @@ import { NavbarComponent } from './navbar/navbar.component';
 import { MatTabsModule } from "@angular/material/tabs";
 import { RouterModule } from "@angular/router";
 import { BrowserAnimationsModule } from "@angular/platform-browser/animations";
-import { MilestonesPageModule } from "./milestones-page/milestones-page.module";
 import { OswegoLogoModule } from "./oswego-logo/oswego-logo.module";
 import { AuthInterceptor } from './security/interceptors/auth-interceptor';
 import { TasksModule } from './tasks/tasks.module';
@@ -81,7 +80,6 @@ import { MilestonesFacultyModule } from './milestones-page/milestones-faculty/mi
     SocialLoginModule,
     DashboardModule,
     PortfolioModule,
-    MilestonesPageModule,
     UsersPageModule,
     MatCardModule,
     MatTabsModule,
