+++ frontend/crd/src/app/milestones-page/milestones-faculty/milestones-faculty.component.ts
@@ -5,7 +5,7 @@ import { takeUntil, zip } from 'rxjs';
 import { MatDialog} from "@angular/material/dialog";
 import { SubmissionService } from 'src/app/submissions/submission.service';
 import { User } from 'src/app/security/domain/user';
-import { MilestonesPageComponent } from '../milestones-page.component';
+import { MilestonesPage } from '../milestones-page';
 import { ActivatedRoute } from '@angular/router';
 import { UserService } from 'src/app/security/user.service';
 
@@ -14,7 +14,7 @@ import { UserService } from 'src/app/security/user.service';
   templateUrl: './milestones-faculty.component.html',
   styleUrls: ['./milestones-faculty.component.less']
 })
-export class MilestonesFacultyComponent extends MilestonesPageComponent implements OnInit, OnDestroy {
+export class MilestonesFacultyComponent extends MilestonesPage implements OnInit, OnDestroy {
 
   studentID!: string;
   currentStudent!: User;
