--- frontend/crd/src/app/milestones-page/milestones-page.component.less
@@ -1,14 +1,0 @@
-@import "../../common";
-
-.page-container {
-  display: flex;
-  align-items: flex-start;
-  flex-direction: column;
-  margin: 0 5vh 5vh 5vh;
-
-  h1 {
-    padding: 6vh 0 6vh 0;
-    font-size: 40px;
-    font-weight: bold;
-  }
-}
